refactor: move Room migrations to KgsDatabaseMigrations
